[Script updates] Added elipsis.css and updated formcheck.js
Elipsis.css will give a super small visual update allowing the user to set a text to a max of 1,2 or 3 lines.

Updated the formcheck script to use classes so the script is better optimized and we can encapsulate the functionalities or trigger them if needed. also added a clear form function that can be triggered through the class or throguh a button if added according to the documentation.